Menu principal enlazado, y login

+++ lib/models/views/balance_general.dart
@@ -9,9 +9,11 @@ class BalanceGeneral {
     required this.balance,
   });
 
-  factory BalanceGeneral.fromMap(Map<String, dynamic> map) => BalanceGeneral(
-    totalIngresos: map['total_ingresos'] ?? 0.0,
-    totalEgresos: map['total_egresos'] ?? 0.0,
-    balance: map['balance'] ?? 0.0,
-  );
+  
+factory BalanceGeneral.fromMap(Map<String, dynamic> map) => BalanceGeneral(
+  totalIngresos: (map['total_ingresos'] ?? 0).toDouble(),
+  totalEgresos: (map['total_egresos'] ?? 0).toDouble(),
+  balance: (map['balance'] ?? 0).toDouble(),
+);
+
 }
